+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- User-function error type
 - TryFrom<> implementations to convert Value into primitive types
 - User Function example and doctest
 - Parse function call for json rules